added auth to header
